Added bilingual news
- Updated DB to include Liams changes
  -> Added french translations for news
- Updated gson to a more current version
- Moved league news from Division to League
