+++ LastEver/src/dao/Division.java
@@ -1,6 +1,5 @@
 package dao;
 
-import db.ConnectionManager;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -9,8 +8,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import beans.DivisionBean;
-import beans.NewsBean;
 import beans.StandingsBean;
+import db.ConnectionManager;
 
 /**
  * The Division class gets all the divisions in the database, allows the user to create a new division, get the news associated
@@ -18,106 +17,6 @@ import beans.StandingsBean;
  */
 public class Division {
 
-	/**
-	 * The getNews method gets all the news associated with the division
-	 * @param <NewsBean>
-	 * @param id - The current division id
-	 * @param lang - The current language of the website
-	 * @param offset - The offset to get the news from (for pagination)
-	 * @param maxRows - The number of news articles to be fetched each time
-	 * @return status - boolean value
-	 */
-	public static boolean getNews(String id, List<NewsBean> news, String lang, int offset, int maxRows) { 
-
-		boolean status = false;					// query status
-		Connection conn = null;					// DB connection
-		PreparedStatement getDivision = null;	// SQL query
-		PreparedStatement getNews = null;		// SQL query
-		ResultSet resultSet = null;				// returned query result set
-		ResultSet rs = null;					// returned query result set
-		String divisionName = null;				// division name to be used in second query
-
-		// Connect to Database and execute SELECT query
-		try {
-			conn = ConnectionManager.getConnection();
-			getDivision = conn.prepareStatement("select divisionName from division where divisionID = ?");
-			getDivision.setString(1, id);
-			
-			resultSet = getDivision.executeQuery();
-			status = resultSet.next();
-
-			//gets the divisionName in order to get the news associated with the division
-			divisionName = resultSet.getString(1);
-
-			//if there is a result then get the news
-			if(status) {
-				getNews = conn.prepareStatement("select u.userName, n.newsTitle, n.newsTime, n.newsContent from news n"
-						+ " inner join users u on u.userID = n.userID inner join newsxtags nt on nt.newsID = n.newsID"
-						+ " inner join tags t on nt.tagID = t.tagID where tagDescription = ? order by n.newsTime desc"
-						+ " LIMIT ?,?");
-				getNews.setString(1, divisionName);
-				getNews.setInt(2, offset);
-				getNews.setInt(3, maxRows);
-				rs = getNews.executeQuery();
-				status = rs.next();
-
-				//return to the start of the result set
-				rs.beforeFirst();
-
-				//Loop through and add the results of the query to a NewsBean then add it to the list
-				while(rs.next()) {
-					NewsBean nb = new NewsBean();
-					nb.setUserName(rs.getString(1));
-					nb.setTitle(rs.getString(2));
-					nb.setPostedTime(rs.getTimestamp(3), lang);
-					nb.setContent(rs.getString(4));
-					news.add(nb);
-				}
-			}
-
-			// close all connections and handle all possible exceptions
-		} catch (Exception e) {
-			System.out.println(e);
-		} finally {
-			if (conn != null) {
-				try {
-					conn.close();
-				} catch (SQLException e) {
-					e.printStackTrace();
-				}
-			}
-			if (getDivision != null) {
-				try {
-					getDivision.close();
-				} catch (SQLException e) {
-					e.printStackTrace();
-				}
-			}
-			if (getNews != null) {
-				try {
-					getNews.close();
-				} catch (SQLException e) {
-					e.printStackTrace();
-				}
-			}
-			if (resultSet != null) {
-				try {
-					resultSet.close();
-				} catch (SQLException e) {
-					e.printStackTrace();
-				}
-			}
-			if (rs != null) {
-				try {
-					rs.close();
-				} catch (SQLException e) {
-					e.printStackTrace();
-				}
-			}
-		}
-		return status;
-	}
-	
 	/**
 	 * The numberOfArticles method gets the number of articles associated with each division
 	 * @param dID - The current division id
