+++ LastEver/src/dao/League.java
@@ -8,6 +8,7 @@ import java.util.List;
 
 import beans.DivisionBean;
 import beans.LeagueBean;
+import beans.NewsBean;
 import db.ConnectionManager;
 
 public class League {
@@ -26,10 +27,10 @@ public class League {
 			league.setString(1, lID);
 			rs = league.executeQuery();
 			status = rs.next();
-			
+
 			//return to the start of the result set
 			rs.beforeFirst();
-			
+
 			//Loop through and add the results of the query to a LeagueBean then add it to the list
 			while(rs.next()) {
 				LeagueBean lb = new LeagueBean();
@@ -39,7 +40,7 @@ public class League {
 				leagueList.add(lb);
 			}
 
-		// close all connections and catch all possible Exceptions
+			// close all connections and catch all possible Exceptions
 		} catch (Exception e) {
 			System.out.println(e);
 		} finally {
@@ -60,7 +61,7 @@ public class League {
 		}	    
 		return status;
 	}
-	
+
 	public static boolean getAllLeagues(List<LeagueBean> leagueList) {
 
 		boolean status = false;							// query status
@@ -74,10 +75,10 @@ public class League {
 			league = conn.prepareStatement("SELECT leagueID, leagueName from league");
 			rs = league.executeQuery();
 			status = rs.next();
-			
+
 			//return to the start of the result set
 			rs.beforeFirst();
-			
+
 			//Loop through and add the results of the query to a LeagueBean then add it to the list
 			while(rs.next()) {
 				LeagueBean lb = new LeagueBean();
@@ -86,7 +87,7 @@ public class League {
 				leagueList.add(lb);
 			}
 
-		// close all connections and catch all possible Exceptions
+			// close all connections and catch all possible Exceptions
 		} catch (Exception e) {
 			System.out.println(e);
 		} finally {
@@ -107,7 +108,7 @@ public class League {
 		}	    
 		return status;
 	}
-	
+
 	public static boolean getLeagueDivisions(String lID, List<DivisionBean> divisionList) {
 
 		boolean status = false;							// query status
@@ -123,10 +124,10 @@ public class League {
 			division.setString(1, lID);
 			rs = division.executeQuery();
 			status = rs.next();
-			
+
 			//return to the start of the result set
 			rs.beforeFirst();
-			
+
 			//Loop through and add the results of the query to a LeagueBean then add it to the list
 			while(rs.next()) {
 				DivisionBean lb = new DivisionBean();
@@ -135,7 +136,7 @@ public class League {
 				divisionList.add(lb);
 			}
 
-		// close all connections and catch all possible Exceptions
+			// close all connections and catch all possible Exceptions
 		} catch (Exception e) {
 			System.out.println(e);
 		} finally {
@@ -156,7 +157,7 @@ public class League {
 		}	    
 		return status;
 	}
-	
+
 	public static boolean getDivisionsByLeague(List<DivisionBean> divs, String leagueId) {
 		boolean status = false;					// Status of createNewUser
 		Connection conn = null;					// DB Connection
@@ -178,7 +179,7 @@ public class League {
 				status = true;
 			}
 
-		// Catch all possible Exceptions
+			// Catch all possible Exceptions
 		} catch (Exception e) {
 			System.out.println(e);
 		} finally {
@@ -197,6 +198,110 @@ public class League {
 				}
 			}
 		}	    
+		return status;
+	}
+
+	/**
+	 * The getNews method gets all the news associated with the league
+	 * @param <NewsBean>
+	 * @param id - The current division id
+	 * @param lang - The current language of the website
+	 * @param offset - The offset to get the news from (for pagination)
+	 * @param maxRows - The number of news articles to be fetched each time
+	 * @return status - boolean value
+	 */
+	public static boolean getNews(String id, List<NewsBean> news, String lang, int offset, int maxRows) { 
+
+		boolean status = false;					// query status
+		Connection conn = null;					// DB connection
+		PreparedStatement getLeague = null;		// SQL query
+		PreparedStatement getNews = null;		// SQL query
+		ResultSet resultSet = null;				// returned query result set
+		ResultSet rs = null;					// returned query result set
+		String leagueName = null;				// division name to be used in second query
+
+		// Connect to Database and execute SELECT query
+		try {
+			conn = ConnectionManager.getConnection();
+			getLeague = conn.prepareStatement("select leagueName from league where leagueID = ?");
+			getLeague.setString(1, id);
+
+			resultSet = getLeague.executeQuery();
+			status = resultSet.next();
+
+			//gets the leagueName in order to get the news associated with the division
+			leagueName = resultSet.getString(1);
+
+			//if there is a result then get the news
+			if(status) {
+				
+				getNews = conn.prepareStatement("select concat_ws(' ', u.userFirstName, u.userLastName), n.newsTitle,"
+						+ " n.newsTitle_fr, n.newsTime, n.newsContent, n.newsContent_fr from news n inner join users"
+						+ " u on u.userID = n.userID inner join newsxtags nt on nt.newsID = n.newsID"
+						+ " inner join tags t on nt.tagID = t.tagID where tagDescription = ? order by n.newsTime desc"
+						+ " LIMIT ?,?");
+				getNews.setString(1, leagueName);
+				getNews.setInt(2, offset);
+				getNews.setInt(3, maxRows);
+				rs = getNews.executeQuery();
+				status = rs.next();
+
+				//return to the start of the result set
+				rs.beforeFirst();
+
+				//Loop through and add the results of the query to a NewsBean then add it to the list
+				while(rs.next()) {
+					NewsBean nb = new NewsBean();
+					nb.setUserName(rs.getString(1));
+					nb.setTitle(rs.getString(2));
+					nb.setTitleFR(rs.getString(3));
+					nb.setPostedTime(rs.getTimestamp(4), lang);
+					nb.setContent(rs.getString(5));
+					nb.setContentFR(rs.getString(6));
+					news.add(nb);
+				}
+			}
+
+			// close all connections and handle all possible exceptions
+		} catch (Exception e) {
+			System.out.println(e);
+		} finally {
+			if (conn != null) {
+				try {
+					conn.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+			if (getLeague != null) {
+				try {
+					getLeague.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+			if (getNews != null) {
+				try {
+					getNews.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+			if (resultSet != null) {
+				try {
+					resultSet.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+			if (rs != null) {
+				try {
+					rs.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+		}
 		return status;
 	}
 }
